fix: backfill failure storm, peer exchange handshake, broadcast-aware peer selection
Three critical fixes:

1. BACKFILL: Consecutive failure skip (50 failures -> skip 10K ticks)
   - Qubic nodes only serve recent ticks; requesting tick 0..59M from a
     node at tick 60M wastes millions of TCP round-trips getting EndResponse
   - After 50 consecutive failures, skip ahead 10K ticks to avoid the storm
   - Changed ticks_failed from AtomicU32 to AtomicU64 to prevent overflow

2. PEER EXCHANGE: Complete the handshake (read node's peer response)
   - Was fire-and-forget: sent peers but never read the response
   - Many nodes won't start broadcasting until handshake completes
   - Now reads type 0 response and adds discovered peers to PeerManager

3. PEER SELECTION: Broadcast-aware scoring
   - Peers that actually broadcast data (type 8, 24, etc.) score 0.7-1.0
   - Non-broadcasting peers cap at 0.5
   - track broadcast_count per peer, prefer broadcasting nodes
   - mark_broadcast() called when broadcast packets arrive in read loop

Also: fixed published_since_last_stats counter (was never incremented)

+++ crates/core/src/pipeline.rs
@@ -66,7 +66,7 @@ pub struct PipelineState {
     /// Total tick data items discovered by the backfill process.
     pub backfill_ticks_discovered: AtomicU64,
     /// Ticks that failed during backfill.
-    pub backfill_ticks_failed: AtomicU32,
+    pub backfill_ticks_failed: AtomicU64,
     /// Backfill start tick.
     pub backfill_start_tick: AtomicU32,
     /// Backfill end tick.
@@ -125,7 +125,7 @@ pub struct PipelineStatusResponse {
     /// Total tick data items discovered by backfill.
     pub backfill_ticks_discovered: u64,
     /// Ticks that failed during backfill.
-    pub backfill_ticks_failed: u32,
+    pub backfill_ticks_failed: u64,
     /// Backfill start tick.
     pub backfill_start_tick: u32,
     /// Backfill end tick.
@@ -162,7 +162,7 @@ impl PipelineState {
             backfill_ticks_completed: AtomicU64::new(0),
             backfill_txs_discovered: AtomicU64::new(0),
             backfill_ticks_discovered: AtomicU64::new(0),
-            backfill_ticks_failed: AtomicU32::new(0),
+            backfill_ticks_failed: AtomicU64::new(0),
             backfill_start_tick: AtomicU32::new(0),
             backfill_end_tick: AtomicU32::new(0),
             started_at: Instant::now(),
